Code Restructuring
- Reset keycloak init method to default configuration
- remove bypass in currentUser and Keycloak check in admin routes
- restructure listFormComponentsHelpInfo from admin to form. This prevent exposing admin store from non admin users
