+++ app/frontend/src/views/form/Create.vue
@@ -58,7 +58,7 @@ import FormDesigner from '@/components/designer/FormDesigner.vue';
 import FormSettings from '@/components/designer/FormSettings.vue';
 import FormDisclaimer from '@/components/designer/FormDisclaimer.vue';
 import { IdentityMode, IdentityProviders } from '@/utils/constants';
-import admin from '@/store/modules/admin.js';
+
 export default {
   name: 'FormCreate',
   components: {
@@ -80,8 +80,7 @@ export default {
     };
   },
   methods: {
-    ...mapActions('admin', ['listFormComponentsHelpInfo']),
-    ...mapActions('form', ['resetForm']),
+    ...mapActions('form', ['listFormComponentsHelpInfo','resetForm']),
     reRenderFormDesigner() {
       this.creatorStep = 2;
       this.$refs.formDesigner.onFormLoad();
@@ -110,18 +109,7 @@ export default {
         )
       )
       : next();
-  },
-  beforeMount(){
-    if(!this.$store.hasModule('admin')) {
-      this.$store.registerModule('admin', admin);
-    }
-  },
-  beforeUnmount(){
-    if (this.$store.hasModule('admin')) {
-      this.$store.unregisterModule('admin');
-    }
-
-  },
+  }
 };
 </script>
