+++ app/src/forms/form/service.js
@@ -15,7 +15,8 @@ const {
   FormSubmissionStatus,
   FormSubmissionUser,
   IdentityProvider,
-  SubmissionMetadata
+  SubmissionMetadata,
+  FormComponentsHelpInfo
 } = require('../common/models');
 const { falsey, queryUtils } = require('../common/utils');
 const { Permissions, Roles, Statuses } = require('../common/constants');
@@ -566,6 +567,27 @@ const service = {
       .throwIfNotFound();
   },
   // ----------------------------------------------------------------------Api Key
+
+  /**
+   * @function listFormComponentsHelpInfo
+   * Search for all form components help information
+   * @returns {Promise} An objection query promise
+   */
+   listFormComponentsHelpInfo: async () => {
+   
+    let result = await FormComponentsHelpInfo.query()
+    .modify('distinctOnComponentName')
+    .modify('orderComponentNameVersionsDescending');
+  
+    let filterResult= result.map(item=>({id:item.id,status:item.publishstatus,componentName:item.componentname,moreHelpInfoLink:item.morehelpinfolink,imageUrl:item.imageurl,
+      version:item.versions,groupName:item.groupname,description:item.description }));
+     
+    return filterResult.reduce(function (r, a) {
+      r[a.groupName] = r[a.groupName] || [];
+      r[a.groupName].push(a);
+      return r;
+    }, Object.create(null));
+  },
 };
 
 module.exports = service;
